refactor: 不要になった generate-holidays.ts を削除し import 拡張子を .ts に統一
ロジックベースの祝日判定に統一したことで、CSV から祝日データを
生成するスクリプトが不要になった。

また、rewriteRelativeImportExtensions の導入により import 拡張子を
.ts に統一した。

- generate-holidays.ts を削除
- テストに必要な関数（parseCsv, fetchCsv 等）を generate-check-all-test.ts に移動
- generateHolidayDatesTs, generateHolidayNamesTs のテストを削除
- package.json から generate スクリプトを削除
- 全ファイルの import 拡張子を .js から .ts に変更
- 開発ルールを更新
