+++ src/isNationalHoliday/index.ts
@@ -1,13 +1,13 @@
-import type { DateInput } from '../types.js';
-import { formatDate } from '../_internal/formatDate.js';
+import type { DateInput } from '../types.ts';
+import { formatDate } from '../_internal/formatDate.ts';
 import {
   getJstFullYear,
   getJstMonth,
   getJstDate,
   getJstDay,
   createJstDate,
-} from '../_internal/jst.js';
-import { addDays } from '../_internal/addDays.js';
+} from '../_internal/jst.ts';
+import { addDays } from '../_internal/addDays.ts';
 
 /**
  * 特別な一回限りの祝日
